+++ src/main/java/module/lineup/ratings/MultipleRatingChartsPanel.java
@@ -26,17 +26,15 @@ public final class MultipleRatingChartsPanel extends JPanel {
 		private final LinesChartDataModel[] data = new LinesChartDataModel[2];
 		private final double[] m_Data;
 		private final double m_AvgValue;
-		private final NumberFormat m_fmt;
 		private final String m_SeriesName;
 		private final boolean m_isType2;
 
-		public SingleChart(double avgValue, double[] values, NumberFormat format, String serieName) {
-			this(avgValue, values, format, serieName, false);
+		public SingleChart(double avgValue, double[] values, String serieName) {
+			this(avgValue, values, serieName, false);
 		}
 
-		public SingleChart(double avgValue, double[] values, NumberFormat format, String serieName, boolean bType2) {
+		public SingleChart(double avgValue, double[] values,  String serieName, boolean bType2) {
 			m_Data = values;
-			m_fmt = format;
 			m_SeriesName = serieName;
 			m_isType2 = bType2;
 			m_AvgValue = avgValue;
@@ -57,8 +55,8 @@ public final class MultipleRatingChartsPanel extends JPanel {
 				lineColor = ThemeManager.getColor(HOColorName.PALETTE13_0);
 			}
 
-			data[0] = new LinesChartDataModel(m_Data, " ", true, lineColor, m_fmt);
-			var data1 = new LinesChartDataModel(avgArray, "avg", true, lineColor, SeriesLines.DASH_DASH, SeriesMarkers.NONE, m_fmt, 1d, false);
+			data[0] = new LinesChartDataModel(m_Data, " ", true, lineColor);
+			var data1 = new LinesChartDataModel(avgArray, "avg", true, lineColor, SeriesLines.DASH_DASH, SeriesMarkers.NONE, 1d, false);
 			data1.setNotVisibleLegend();
 			data[1] = data1;
 
@@ -69,10 +67,6 @@ public final class MultipleRatingChartsPanel extends JPanel {
 
 		public void setHelpLines(boolean state) {
 			oChartPanel.setHelpLines(state);
-		}
-
-		public void setValues(boolean state) {
-			oChartPanel.setLabelling(state);
 		}
 
 		public JPanel getChartPanel() {
@@ -116,41 +110,41 @@ public final class MultipleRatingChartsPanel extends JPanel {
 		gbc.weightx = 1;
 		gbc.weighty = 1;
 		
-		leftDefense = new SingleChart(chartData.getLeftDefenceAvg(), chartData.getLeftDefence(),fmt , Helper.getTranslation("ls.match.ratingsector.leftdefence"));
+		leftDefense = new SingleChart(chartData.getLeftDefenceAvg(), chartData.getLeftDefence(), Helper.getTranslation("ls.match.ratingsector.leftdefence"));
 		chartsPanel.add(leftDefense.getChartPanel(), gbc);
 
 		gbc.gridx = 1;
-		centralDefense = new SingleChart(chartData.getCentralDefenceAvg(), chartData.getCentralDefence(),	fmt, Helper.getTranslation("ls.match.ratingsector.centraldefence"));
+		centralDefense = new SingleChart(chartData.getCentralDefenceAvg(), chartData.getCentralDefence(), Helper.getTranslation("ls.match.ratingsector.centraldefence"));
 		chartsPanel.add(centralDefense.getChartPanel(), gbc);
 
 		gbc.gridx = 2;
-		rightDefense = new SingleChart(chartData.getRightDefenceAvg(), chartData.getRightDefence(),	Helper.DEFAULTDEZIMALFORMAT, Helper.getTranslation("ls.match.ratingsector.rightdefence"));
+		rightDefense = new SingleChart(chartData.getRightDefenceAvg(), chartData.getRightDefence(), Helper.getTranslation("ls.match.ratingsector.rightdefence"));
 		chartsPanel.add(rightDefense.getChartPanel(), gbc);
 
 		gbc.gridx = 0;
 		gbc.gridy = 1;
-		hatStats = new SingleChart(chartData.getHatStatsAvg(), chartData.getHatStats(),	Helper.INTEGERFORMAT, Helper.getTranslation("ls.match.ratingtype.hatstats"), true);
+		hatStats = new SingleChart(chartData.getHatStatsAvg(), chartData.getHatStats(), Helper.getTranslation("ls.match.ratingtype.hatstats"), true);
 		chartsPanel.add(hatStats.getChartPanel(), gbc);
 
 		gbc.gridx = 1;
-		midfield = new SingleChart(chartData.getMidfieldAvg(), chartData.getMidfield(),	fmt, Helper.getTranslation("ls.match.ratingsector.midfield"));
+		midfield = new SingleChart(chartData.getMidfieldAvg(), chartData.getMidfield(), Helper.getTranslation("ls.match.ratingsector.midfield"));
 		chartsPanel.add(midfield.getChartPanel(), gbc);
 
 		gbc.gridx = 2;
-		loddar = new SingleChart(chartData.getLoddarAvg(), chartData.getLoddar(), fmt, Helper.getTranslation("ls.match.ratingtype.loddarstats"), true);
+		loddar = new SingleChart(chartData.getLoddarAvg(), chartData.getLoddar(), Helper.getTranslation("ls.match.ratingtype.loddarstats"), true);
 		chartsPanel.add(loddar.getChartPanel(), gbc);
 
 		gbc.gridx = 0;
 		gbc.gridy = 2;
-		leftAttack = new SingleChart(chartData.getLeftAttackAvg(), chartData.getLeftAttack(),	fmt, Helper.getTranslation("ls.match.ratingsector.leftattack"));
+		leftAttack = new SingleChart(chartData.getLeftAttackAvg(), chartData.getLeftAttack(), Helper.getTranslation("ls.match.ratingsector.leftattack"));
 		chartsPanel.add(leftAttack.getChartPanel(), gbc);
 
 		gbc.gridx = 1;
-		centralAttack = new SingleChart(chartData.getCentralAttackAvg(), chartData.getCentralAttack(), fmt, Helper.getTranslation("ls.match.ratingsector.centralattack"));
+		centralAttack = new SingleChart(chartData.getCentralAttackAvg(), chartData.getCentralAttack(), Helper.getTranslation("ls.match.ratingsector.centralattack"));
 		chartsPanel.add(centralAttack.getChartPanel(), gbc);
 
 		gbc.gridx = 2;
-		rightAttack = new SingleChart(chartData.getRightAttackAvg(), chartData.getRightAttack(), fmt,	Helper.getTranslation("ls.match.ratingsector.rightattack"));
+		rightAttack = new SingleChart(chartData.getRightAttackAvg(), chartData.getRightAttack(),	Helper.getTranslation("ls.match.ratingsector.rightattack"));
 		chartsPanel.add(rightAttack.getChartPanel(), gbc);
 	}